Crawl facebook is its own file
